Reduce breakability of deleted reviews.
Fixes #4
Log the last 60 reviews (more than a single page of iOS reviews) and
check each instance of a logged review to ensure that a deleted review
won't cause a repost of the previous (valid) review. Does not have
protection for a user posting multiple times.

Fix for rate-limited translations from API.